Upgrade for qio version 2 Memory economization Malloc tracking

+++ generic/layout_hyper_prime.c
@@ -123,7 +123,7 @@ register int i,xr,yr,zr,tr;
     }
 }
 
-int num_sites(int node) {
+size_t num_sites(int node) {
     return( sites_on_node );
 }
