+++ generic/layout_timeslices.c
@@ -71,6 +71,6 @@ register int i, tr;
     }
 }
 
-int num_sites(int node) {
+size_t num_sites(int node) {
     return( sites_on_node );
 }
